Started working to make the listing.blade.php dynamic
Made the 'grid' list the default. If the entered URL in the browser address bar is a 'category', we need to fetch its related products as well as its subcategories related products, but if the URL is a subcategory, we need to fetch the subcategory related products only. Created categoryDetails() method in Category.php model